requirements.txt
fastapi==0.115.5
langchain_core==0.3.19
langchain_groq==0.2.1
pydantic==2.9.2
python-dotenv==1.0.1
pytest
